+++ src/components/FriendSelector.tsx
@@ -1,7 +1,7 @@
 /** @jsxImportSource @emotion/react */
 import styled from '@emotion/styled';
 import { FaPlus } from 'react-icons/fa';
-import { UserManagement } from '../pages/Login/UserManagement';
+import { UserManagement } from '../pages/Login/contexts/UserManagement';
 
 const Wrapper = styled.div`
   background-color: ${({ theme }) => theme.color.gray.gray300};
