+++ src/pages/MyPage.tsx
@@ -1,6 +1,6 @@
 /** @jsxImportSource @emotion/react */
 import { useTheme } from '@emotion/react';
-import { UserManagement } from './Login/UserManagement';
+import { UserManagement } from './Login/contexts/UserManagement';
 import { css } from '@emotion/react';
 import { useNavigate } from 'react-router-dom';
